requirements.txt
django==3.2
djangorestframework==3.11
pandas==1.2.1